requirements/default.txt
git+https://github.com/acrcloud/acrcloud_sdk_python
scp
paramiko
python-dotenv